requirements.txt
flask
anthropic
numpy
python-dotenv
sentence-transformers